Extended the baseCell style for th and td with specific styles.
Extended the baseCell style for th and td with specific styles using the array syntax provided by @vanilla-extract/css. This way, we can add or override styles specific to th and td while still keeping the base cell styling intact.

Also to note, this refactor reduces repetition, makes the code more maintainable, and keeps it scalable for future additions or changes to the table cell styling.